+++ packages/subgraph/subgraph.template.yaml
@@ -104,7 +104,7 @@ dataSources:
         kind: ethereum/events
         apiVersion: 0.0.4
         language: wasm/assemblyscript
-        file: ./src/mappings/MmaMarketFactory.ts
+        file: ./src/mappings/MmaMarketFactory{{ subtype }}.ts
         entities:
           - MarketCreated
           - MarketResolved
